requirements-dev.txt
black
pytest
setuptools
requests_mock